requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10

# LLM
ollama>=0.1.0

# Memory - Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Voice - Speech to Text
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.0

# Voice - Text to Speech
pyttsx3>=2.90

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
